Improved soundness under "Stacked Borrows" model
The conversation handler is no longer kept in a `Box` directly in
`Context`. We always had to hand a pointer to it to PAM C code, so
we now use `Box::into_raw()` and `Box::from_raw()` on that pointer
as needed.

This makes `Context` a bit smaller (now one pointer + one `int`)
and there is no possibility of invalidated pointers when
roundtripping through C anymore. As a small performance downside
`Context::conversation()` and `Context::conversation_mut()` now
always involve a function call that cannot be optimized away.